deleted unnecessary user routes, replaced auth routes with users routes to align with our documentation
