+++ app/__init__.py
@@ -7,7 +7,6 @@ from flask_login import LoginManager
 
 from .models import db, User
 from .routes.user_routes import user_routes
-from .routes.auth_routes import auth_routes
 
 from .seeds import seed_commands
 
@@ -29,8 +28,7 @@ def load_user(id):
 app.cli.add_command(seed_commands)
 
 app.config.from_object(Config)
-app.register_blueprint(user_routes, url_prefix='/api/users')
-app.register_blueprint(auth_routes, url_prefix='/api/auth')
+app.register_blueprint(user_routes, url_prefix='/users')
 db.init_app(app)
 Migrate(app, db)
