+++ react-app/src/services/auth.js
@@ -1,38 +1,37 @@
-export const authenticate = async() => {
-  const response = await fetch('/api/auth/',{
-    headers: {
-      'Content-Type': 'application/json'
-    }
-  });
-  return await response.json();
-}
-
-export const login = async (email, password) => {
-  const response = await fetch('/api/auth/login', {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json'
-    },
-    body: JSON.stringify({
-      email,
-      password
-    })
-  });
-  return await response.json();
-}
-
-export const logout = async () => {
-  const response = await fetch("/api/auth/logout", {
+export const authenticate = async () => {
+  const response = await fetch("/users/", {
     headers: {
       "Content-Type": "application/json",
-    }
+    },
   });
   return await response.json();
 };
 
+export const login = async (email, password) => {
+  const response = await fetch("/users/login", {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify({
+      email,
+      password,
+    }),
+  });
+  return await response.json();
+};
+
+export const logout = async () => {
+  const response = await fetch("/users/logout", {
+    headers: {
+      "Content-Type": "application/json",
+    },
+  });
+  return await response.json();
+};
 
 export const signUp = async (username, email, password) => {
-  const response = await fetch("/api/auth/signup", {
+  const response = await fetch("/users/signup", {
     method: "POST",
     headers: {
       "Content-Type": "application/json",
@@ -44,4 +43,4 @@ export const signUp = async (username, email, password) => {
     }),
   });
   return await response.json();
-}
+};
